set up glbtile/glbtop rtl scripts to check if input rtl is already available

+++ mflowgen/glb_tile/rtl/gen_rtl.sh
@@ -1,11 +1,17 @@
 #!/bin/bash
-
-# SystemRDL run
-make -C $GARNET_HOME/global_buffer rtl
-
-rm -f outputs/design.v
-
-while read F  ; do
-    echo "Reading design file: $F"
-    cat $GARNET_HOME/global_buffer/rtl/$F >> outputs/design.v
-done <$GARNET_HOME/global_buffer/rtl/glb_tile.filelist
+# Glb can accept rtl as an input from parent graph
+if [ -f ../inputs/design.v ]; then
+  echo "Using RTL from parent graph"
+  mkdir -p outputs
+  (cd outputs; ln -s ../../inputs/design.v)
+else
+  # SystemRDL run
+  make -C $GARNET_HOME/global_buffer rtl
+  
+  rm -f outputs/design.v
+  
+  while read F  ; do
+      echo "Reading design file: $F"
+      cat $GARNET_HOME/global_buffer/rtl/$F >> outputs/design.v
+  done <$GARNET_HOME/global_buffer/rtl/glb_tile.filelist
+fi
